requirements.txt
langchain
langchain-core
langchain-google-genai
python-dotenv
streamlit
torch
torchvision
pydantic
transformers